+++ index.js
@@ -1,0 +1,7 @@
+panel.plugin('jan-herman/simple-taxonomies', {
+    fields: {
+        'synced-structure': {
+            extends: 'k-structure-field',
+        },
+    },
+});