+++ index.php
@@ -4,11 +4,12 @@ use Kirby\Cms\App as Kirby;
 use Kirby\Cms\Page;
 use Kirby\Cms\Structure;
 use Kirby\Cms\StructureObject;
+use Kirby\Data\Yaml;
+use Kirby\Exception\Exception;
 
 Kirby::plugin('jan-herman/simple-taxonomies', [
     'pagesMethods' => [
-        'taxonomyTerms' => function (string $taxonomy = 'categories'): Structure
-        {
+        'taxonomyTerms' => function (string $taxonomy = 'categories'): Structure {
             $terms = new Structure();
 
             foreach ($this as $page) {
@@ -25,16 +26,13 @@ Kirby::plugin('jan-herman/simple-taxonomies', [
         }
     ],
     'pageMethods' => [
-        'taxonomyArchiveSlug' => function (string $taxonomy = 'categories'): string
-        {
+        'taxonomyArchiveSlug' => function (string $taxonomy = 'categories'): string {
             return (string) param($this->{$taxonomy . '_slug'}()->toString() ?: 'category');
         },
-        'isTaxonomyArchive' => function (string $taxonomy = 'categories'): bool
-        {
+        'isTaxonomyArchive' => function (string $taxonomy = 'categories'): bool {
             return (bool) $this->taxonomyArchiveSlug($taxonomy);
         },
-        'taxonomyTerms' => function (string $taxonomy = 'categories'): Structure
-        {
+        'taxonomyTerms' => function (string $taxonomy = 'categories'): Structure {
             $terms = $this?->content()->{$taxonomy}()?->toStructure();
 
             if (!$terms) {
@@ -51,36 +49,48 @@ Kirby::plugin('jan-herman/simple-taxonomies', [
 
             return $terms;
         },
-        'taxonomyTerm' => function (string $term_slug, string $taxonomy = 'categories'): StructureObject
-        {
-            return $this->taxonomyTerms($taxonomy)->findBy('slug', $term_slug);
+        'taxonomyTerm' => function (string $term_uuid, string $taxonomy = 'categories'): StructureObject {
+            return $this->taxonomyTerms($taxonomy)->findBy('uuid', $term_uuid);
         },
-        'terms' => function (string $taxonomy = 'categories', Page $taxonomy_page = null): Structure
-        {
+        'terms' => function (string $taxonomy = 'categories', Page $taxonomy_page = null): Structure {
             $taxonomy_page = $taxonomy_page ?: $this->parent();
             $taxonomy_terms = $taxonomy_page->taxonomyTerms($taxonomy);
-            $slugs = $this->content()->{$taxonomy}()?->split();
+            $uuids = $this->content()->{$taxonomy}()?->split();
 
-            if (!$taxonomy_terms || !$slugs) {
+            if (!$taxonomy_terms || !$uuids) {
                 return new Structure();
             }
 
-            return $taxonomy_terms->filterBy('slug', 'in', $slugs);
+            return $taxonomy_terms->filterBy('uuid', 'in', $uuids);
         }
     ],
     'fieldMethods' => [
-        'toTerms' => function ($field, Page $taxonomy_page = null): Structure
-        {
+        'toTerms' => function ($field, Page $taxonomy_page = null): Structure {
             $taxonomy_page = $taxonomy_page ?: $field->parent()->parent();
             $taxonomy_terms = $taxonomy_page->taxonomyTerms($field->key());
-            $slugs = $field->split();
+            $uuids = $field->split();
 
-            if (!$taxonomy_terms || !$slugs) {
+            if (!$taxonomy_terms || !$uuids) {
                 return new Structure();
             }
 
-            return $taxonomy_terms->filterBy('slug', 'in', $slugs);
+            return $taxonomy_terms->filterBy('uuid', 'in', $uuids);
         }
+    ],
+    'validators' => [
+        'unique' => function ($value, $field) {
+            $values = array_column(Yaml::decode($value), $field);
+            $is_valid = count($values) === count(array_flip($values));
+
+            if (!$is_valid) {
+                throw new Exception(tt('jan-herman.simple-taxonomies.validators.unique.exception', ['field' => $field]));
+            }
+
+            return $is_valid;
+        }
+    ],
+    'fields' => [
+        'synced-structure' => require_once 'synced-structure-field/synced-structure.php',
     ],
     'blueprints' => [
         'fields/taxonomy'       => __DIR__ . '/blueprints/fields/taxonomy.yml',
@@ -89,8 +99,10 @@ Kirby::plugin('jan-herman/simple-taxonomies', [
     ],
     'translations' => [
         'en' => [
+            'jan-herman.simple-taxonomies.validators.unique.exception' => 'Error: Field \'{ field }\' must be unique.',
             'jan-herman.simple-taxonomies.fields.taxonomy-terms.label' => 'Categories',
             'jan-herman.simple-taxonomies.fields.taxonomy-slug.label' => 'Category Archive URL',
+            'jan-herman.simple-taxonomies.fields.taxonomy-slug.after' => ':category-slug',
             'jan-herman.simple-taxonomies.fields.taxonomy.label' => 'Categories',
             'jan-herman.simple-taxonomies.fields.taxonomy.empty' => 'You haven\'t created any categories yet.',
             'jan-herman.simple-taxonomies.fields.taxonomy.fields.title.label' => 'Title',
@@ -98,8 +110,10 @@ Kirby::plugin('jan-herman/simple-taxonomies', [
             'jan-herman.simple-taxonomies.fields.taxonomy.fields.singular_title.label' => 'Singular Title',
         ],
         'cs' => [
+            'jan-herman.simple-taxonomies.validators.unique.exception' => 'Chyba: Pole \'{ field }\' musí být unikátní.',
             'jan-herman.simple-taxonomies.fields.taxonomy-terms.label' => 'Kategorie',
             'jan-herman.simple-taxonomies.fields.taxonomy-slug.label' => 'URL archivu kategorie',
+            'jan-herman.simple-taxonomies.fields.taxonomy-slug.after' => ':nazev-kategorie',
             'jan-herman.simple-taxonomies.fields.taxonomy.label' => 'Kategorie',
             'jan-herman.simple-taxonomies.fields.taxonomy.empty' => 'Zatím jste nevytvořili žádné kategorie.',
             'jan-herman.simple-taxonomies.fields.taxonomy.fields.title.label' => 'Název',