+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # Changelog
 
+## [2.0.0] - 2024-01-19
+### Added
+- hidden uuid field
+- slug validation (ensures the slug is unique)
+
+### Changed
+- plugin now saves category uuid instead of slug
+- classic structure field was replaced by `synced-structure-field`
+
+### Fixed
+- Missing translations
+
+
 ## [1.0.0] - 2023-02-20
 ### Added
 - Initial release